requirements.txt
plotly
dash
pdf2image
poppler-utils
pytesseract
pillow
PyPDF3
watchdog
torch
transformers==3.1.0
awca